ms-embedding-documents/requirements.txt
azure-search-documents==11.5.2
PyPDF2==3.0.1
python-dotenv==1.0.1